assets/2020-01-18/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.3
beautifulsoup4==4.8.2
django-bootstrap4==1.1.1
django-extensions==2.2.6
django==3.0.2
grpcio-tools==1.26.0
grpcio==1.26.0
gunicorn==20.0.4
opentelemetry-api==0.3a0
opentelemetry-ext-wsgi==0.3a0
opentelemetry-sdk==0.3a0
protobuf==3.11.2
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
